# docs/OPENAPI_COMMERCE.yaml
openapi: 3.0.3
info:
  title: Voucher Commerce Integration API
  version: 1.0.0
  description: Platform-agnostic discount validation and redemption confirmation.
servers:
  - url: https://your-domain.com
paths:
  /api/commerce/discount/validate:
    post:
      summary: Validate a discount code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DiscountValidateRequest'
      responses:
        '200':
          description: Validation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiscountValidateResponse'
...
components:
  schemas:
    Metadata:
      type: object
      properties:
        platform:
          type: string
          enum: [shopify, woocommerce, magento, custom]
        shopId:
          type: string
        orderId:
          type: string
        customerId:
          type: string
        refSource:
          type: string
          enum: [email, qr, share]
        refParam:
          type: string
        couponCode:
          type: string
    DiscountValidateRequest:
      type: object
      required: [code, orderTotal, currency]
      properties:
        code:
          type: string
        orderTotal:
          type: integer
          description: Minor units (cents)
        currency:
          type: string
        customerEmail:
          type: string
        referralId:
          type: string
        metadata:
          $ref: '#/components/schemas/Metadata'
    DiscountValidateResponse:
      type: object
      required: [valid]
      properties:
        valid:
          type: boolean
        discountAmount:
          type: integer
        voucherId:
          type: string
        referralId:
          type: string
        reason:
          type: string
...